add more tests

+++ contracts/OptionToken.sol
@@ -1,56 +1,69 @@
 pragma solidity 0.5.17;
 
-import './libraries/IERC20.sol';
-import './libraries/SafeMath.sol';
+import "./libraries/IERC20.sol";
+import "./libraries/SafeMath.sol";
 
 contract OptionToken is IERC20 {
-    
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
-    string public constant name = 'Option Token';
-    string public constant symbol = 'OPT';
-    uint8  public decimals;
-    uint   public totalSupply;
+    string public constant name = "Option Token";
+    string public constant symbol = "OPT";
+    uint8 public decimals;
+    uint256 public totalSupply;
 
-    mapping(address => uint) public balanceOf;
-    mapping(address => mapping(address => uint)) public allowance;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
-    function _mint(address to, uint value) internal {
+    function _mint(address to, uint256 value) internal {
         totalSupply = totalSupply.add(value);
         balanceOf[to] = balanceOf[to].add(value);
         emit Transfer(address(0), to, value);
     }
 
-    function _burn(address from, uint value) internal {
+    function _burn(address from, uint256 value) internal {
         balanceOf[from] = balanceOf[from].sub(value);
         totalSupply = totalSupply.sub(value);
         emit Transfer(from, address(0), value);
     }
 
-    function _approve(address owner, address spender, uint value) private {
+    function _approve(
+        address owner,
+        address spender,
+        uint256 value
+    ) private {
         allowance[owner][spender] = value;
         emit Approval(owner, spender, value);
     }
 
-    function _transfer(address from, address to, uint value) internal {
+    function _transfer(
+        address from,
+        address to,
+        uint256 value
+    ) internal {
         balanceOf[from] = balanceOf[from].sub(value);
         balanceOf[to] = balanceOf[to].add(value);
         emit Transfer(from, to, value);
     }
 
-    function approve(address spender, uint value) external returns (bool) {
+    function approve(address spender, uint256 value) external returns (bool) {
         _approve(msg.sender, spender, value);
         return true;
     }
 
-    function transfer(address to, uint value) external returns (bool) {
+    function transfer(address to, uint256 value) external returns (bool) {
         _transfer(msg.sender, to, value);
         return true;
     }
 
-    function transferFrom(address from, address to, uint value) external returns (bool) {
-        if (allowance[from][msg.sender] != uint(-1)) {
-            allowance[from][msg.sender] = allowance[from][msg.sender].sub(value);
+    function transferFrom(
+        address from,
+        address to,
+        uint256 value
+    ) external returns (bool) {
+        if (allowance[from][msg.sender] != uint256(-1)) {
+            allowance[from][msg.sender] = allowance[from][msg.sender].sub(
+                value
+            );
         }
         _transfer(from, to, value);
         return true;
